+++ Sources/KeyboardKit/Layout/KeyboardLayoutService.swift
@@ -27,3 +27,38 @@ public protocol KeyboardLayoutService: AnyObject {
         for context: KeyboardContext
     ) -> KeyboardLayout
 }
+
+public extension KeyboardLayout {
+
+    /// This error can be thrown by ``KeyboardLayoutService/tryRegisterLocalizedLayoutService(_:)``.
+    enum TryRegisterLocalizedLayoutServiceError: Error {
+
+        /// The service doesn't support localized layout service registrations.
+        case serviceDoesNotSupportLocalizedServiceRegistration
+
+        /// The provided service doesn't implement the ``LocalizedService`` protocol.
+        case providedServiceDoesNotImplementLocalizedServiceProtocol
+    }
+}
+
+public extension KeyboardLayoutService {
+
+    /// Try to register a localized layout service.
+    ///
+    /// - Parameters:
+    ///   - service: The service to register.
+    ///
+    /// - Throws: ``KeyboardLayout/TryRegisterLocalizedLayoutServiceError``
+    /// if the current service can't be cast to a ``KeyboardLayout/StandardService``
+    /// or the provided `service` doesn't implement ``LocalizedService``.
+    func tryRegisterLocalizedService(
+        _ service: KeyboardLayoutService
+    ) throws {
+        typealias ErrorType = KeyboardLayout.TryRegisterLocalizedLayoutServiceError
+        let selfError = ErrorType.serviceDoesNotSupportLocalizedServiceRegistration
+        let serviceError = ErrorType.providedServiceDoesNotImplementLocalizedServiceProtocol
+        guard let _self = self as? KeyboardLayout.StandardService else { throw selfError }
+        guard let service = service as? KeyboardLayoutService & LocalizedService else { throw serviceError }
+        _self.registerLocalizedService(service)
+    }
+}